Add Wifi connection and service advertisement
This adds connection to a wifi network, as well as advertisement of the
light pet client service via UDP multicast.

The wifi credentials for the network are loaded via the
wifi.credentials.h file, where you define WIFI_SSID and WIFI_PASSWORD.
I have added this file to the gitignore to avoid leaking sensitive ssid
and password information.