Add reaction to server advertisement
This adds preliminary handling of the server advertisement message.

At the moment we just look for incoming UDP packets on the right port.
If they have the special server magic string, we then set what will
become our target IP for data packets to the IP that the message came
from. If they don't but they're on the correct port, we just print the
data we got from them for the moment.