Sends the unix time based on NTP in protobuf
We now send the current unix time in the encoded protobuf message. The
NTP time is grabbed at setup (in the future this will be grabbed at regular
intervals) and we compute the difference between the millis when we
grabbed the time and the current millis, then add this to the NTP datum
to get the current unix time.