Rename src folder to clientSrc
This is needed because we're now going to store both client and server
code in this repo. This will save us having to reference files from a
separate repo when I factor out some of the magic strings and port
numbers etc. into configuration options. These configuration options
can reside in a common directory and the code can pull what it needs
from there.