Add periodic NTP sync and fix timestamp format
This turns the NTP sync into a task that is run by our scheduler. At
the moment we resync with NTP time every hour. We'll see if this is
enough to prevent clock drift once we actually start logging data.

This also fixes what we're sending over the wire as our timestamp.
We're now correctly sending unix time in ms, which is then converted to
seconds on the server end for conversion to a datetime object.
Something was going wrong here, and I suspect it might have been 32bit
integer overflow, hence why all of the relevant variables related to
the NTP time calculations are now unit64_t or cast to it before
calculation.