Convert protobuf types to lists in properties
The internal protobuf stores data in weird objects. This converts these
to lists for the audio and light data. We were already doing this
implicitly using the list comprehension used to scale the temperature
and humidity.

This is required so that we can convert the object's data to JSON for
logging.