Add logging of data and status
This adds logging of information such as advertisements and logging of
data to a file in json format.

In order to separate the data from the information about advertisments,
we set an attribute on the data LogRecords that indicates that they are
data (record.data=True) and then filter these at the file handler that
logs this data.